+++ back/src/auth/jwt.strategy.ts
@@ -18,10 +18,10 @@ export class JwtStrategy extends PassportStrategy(Strategy) {
     });
   }
 
-  async validate(req: any, payload: any) {
+  async validate(req: Request, payload: { sub: number; iat: number }): Promise<{ userId: number }> {
     const token = ExtractJwt.fromAuthHeaderAsBearerToken()(req);
     
-    if (this.authService.isTokenBlacklisted(token)) {
+    if (await this.authService.isTokenBlacklisted(token)) {
       throw new UnauthorizedException('Token has been invalidated');
     }